# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateSubflagRequest:
    properties:
      name:
        type: string
      sortOrder:
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      requestId:
        type: string
    type: object
...
  dto.FlagObject:
    properties:
      color:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.ListSubflagsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.SubflagResponse'
        type: array
      nextCursor:
        type: string
    type: object
...
  dto.SubflagResponse:
    properties:
      color:
        type: string
      createdAt:
        type: string
      flag:
        $ref: '#/definitions/dto.FlagObject'
      id:
        type: string
      name:
        type: string
      sortOrder:
        type: integer
      updatedAt:
        type: string
    type: object
...
  dto.UpdateSubflagRequest:
    properties:
      name:
        type: string
      sortOrder:
        type: integer
    type: object
...
info:
  contact: {}
  description: API do MVP Organiq.
  title: Organiq API
  version: "0.1"
paths:
...
  /v1/flags/{id}/subflags:
    get:
      parameters:
      - description: Flag ID
        in: path
        name: id
        required: true
        type: string
      - description: Limite
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListSubflagsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar subflags por flag
      tags:
      - Subflags
    post:
      consumes:
      - application/json
      parameters:
      - description: Flag ID
        in: path
        name: id
        required: true
        type: string
      - description: Subflag payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubflagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SubflagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Criar subflag
      tags:
      - Subflags
...
  /v1/subflags/{id}:
    delete:
      parameters:
      - description: Subflag ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remover subflag
      tags:
      - Subflags
    patch:
      consumes:
      - application/json
      parameters:
      - description: Subflag ID
        in: path
        name: id
        required: true
        type: string
      - description: Subflag payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubflagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SubflagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Atualizar subflag
      tags:
      - Subflags
...
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"